Fix hnh dashboard view
